msm: kgsl: Clear and set device started bit on soft reset
When device is soft reset then the started flag which indicates that
device is inactive needs to be reset and set again when the reset
completes. This prevents the code from hitting a BUG while waiting
for dispatcher to idle when the device is being reset from the
dispatcher code.

Change-Id: I669f5e110585f1a934e919a3a804684997f5d82f
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>

Conflicts:
	drivers/gpu/msm/adreno.c